refactor course page, move course-header to its own file
