+++ apps/main/messages/en.po
@@ -134,20 +134,10 @@ msgctxt "R/6nsx"
 msgid "Subscription"
 msgstr "Subscription"
 
-#: src/app/[locale]/(catalog)/b/[brandSlug]/c/[courseSlug]/page.tsx
-msgctxt "cvtZ2C"
-msgid "Course not found"
-msgstr "Course not found"
-
-#: src/app/[locale]/(catalog)/b/[brandSlug]/c/[courseSlug]/page.tsx
+#: src/app/[locale]/(catalog)/b/[brandSlug]/c/[courseSlug]/course-header.tsx
 msgctxt "Dwdx/1"
 msgid "This content was generated by AI. It may contain errors or inaccuracies and should not be considered professional advice."
 msgstr "This content was generated by AI. It may contain errors or inaccuracies and should not be considered professional advice."
-
-#: src/app/[locale]/(catalog)/b/[brandSlug]/c/[courseSlug]/page.tsx
-msgctxt "oCQePC"
-msgid "Learn with interactive lessons and activities. Explore this course and start your learning journey."
-msgstr "Learn with interactive lessons and activities. Explore this course and start your learning journey."
 
 #: src/app/[locale]/(catalog)/courses/category-pills.tsx
 msgctxt "tuZsq+"
