+++ apps/main/messages/es.po
@@ -134,20 +134,10 @@ msgctxt "R/6nsx"
 msgid "Subscription"
 msgstr "Suscripción"
 
-#: src/app/[locale]/(catalog)/b/[brandSlug]/c/[courseSlug]/page.tsx
-msgctxt "cvtZ2C"
-msgid "Course not found"
-msgstr "Curso no encontrado"
-
-#: src/app/[locale]/(catalog)/b/[brandSlug]/c/[courseSlug]/page.tsx
+#: src/app/[locale]/(catalog)/b/[brandSlug]/c/[courseSlug]/course-header.tsx
 msgctxt "Dwdx/1"
 msgid "This content was generated by AI. It may contain errors or inaccuracies and should not be considered professional advice."
 msgstr "Este contenido fue generado por IA. Puede contener errores o imprecisiones y no debe considerarse asesoramiento profesional."
-
-#: src/app/[locale]/(catalog)/b/[brandSlug]/c/[courseSlug]/page.tsx
-msgctxt "oCQePC"
-msgid "Learn with interactive lessons and activities. Explore this course and start your learning journey."
-msgstr "Aprende con lecciones y actividades interactivas. Explora este curso y comienza tu viaje de aprendizaje."
 
 #: src/app/[locale]/(catalog)/courses/category-pills.tsx
 msgctxt "tuZsq+"
