+++ apps/main/messages/pt.po
@@ -134,20 +134,10 @@ msgctxt "R/6nsx"
 msgid "Subscription"
 msgstr "Assinatura"
 
-#: src/app/[locale]/(catalog)/b/[brandSlug]/c/[courseSlug]/page.tsx
-msgctxt "cvtZ2C"
-msgid "Course not found"
-msgstr "Curso não encontrado"
-
-#: src/app/[locale]/(catalog)/b/[brandSlug]/c/[courseSlug]/page.tsx
+#: src/app/[locale]/(catalog)/b/[brandSlug]/c/[courseSlug]/course-header.tsx
 msgctxt "Dwdx/1"
 msgid "This content was generated by AI. It may contain errors or inaccuracies and should not be considered professional advice."
 msgstr "Este conteúdo foi gerado por IA. Pode conter erros ou imprecisões e não deve ser considerado aconselhamento profissional."
-
-#: src/app/[locale]/(catalog)/b/[brandSlug]/c/[courseSlug]/page.tsx
-msgctxt "oCQePC"
-msgid "Learn with interactive lessons and activities. Explore this course and start your learning journey."
-msgstr "Aprenda com aulas e atividades interativas. Explore este curso e comece sua jornada de aprendizado."
 
 #: src/app/[locale]/(catalog)/courses/category-pills.tsx
 msgctxt "tuZsq+"
