+++ apps/main/src/app/[locale]/(catalog)/b/[brandSlug]/c/[courseSlug]/course-header.tsx
@@ -1,0 +1,109 @@
+import { Badge } from "@zoonk/ui/components/badge";
+import {
+  MediaCard,
+  MediaCardContent,
+  MediaCardDescription,
+  MediaCardHeader,
+  MediaCardIcon,
+  MediaCardImage,
+  MediaCardIndicator,
+  MediaCardPopover,
+  MediaCardPopoverAIWarning,
+  MediaCardPopoverBadges,
+  MediaCardPopoverMeta,
+  MediaCardPopoverSource,
+  MediaCardPopoverText,
+  MediaCardTitle,
+  MediaCardTrigger,
+} from "@zoonk/ui/components/media-card";
+import { AI_ORG_SLUG } from "@zoonk/utils/constants";
+import { NotebookPenIcon } from "lucide-react";
+import Image from "next/image";
+import { getExtracted, getLocale } from "next-intl/server";
+import type { CourseWithDetails } from "@/data/courses/get-course";
+import { Link } from "@/i18n/navigation";
+import { getCategories } from "@/lib/categories";
+
+export async function CourseHeader({
+  brandSlug,
+  course,
+}: {
+  brandSlug: string;
+  course: CourseWithDetails;
+}) {
+  const t = await getExtracted();
+  const locale = await getLocale();
+  const categoryLabels = await getCategories({ locale });
+  const courseCategoryKeys = course.categories.map((c) => c.category);
+
+  const displayCategories = categoryLabels.filter((cat) =>
+    courseCategoryKeys.includes(cat.key),
+  );
+
+  const isAIGenerated = brandSlug === AI_ORG_SLUG;
+
+  return (
+    <MediaCard>
+      <MediaCardTrigger>
+        {course.imageUrl ? (
+          <MediaCardImage>
+            <Image
+              alt={course.title}
+              className="size-full object-cover"
+              fill
+              sizes="(max-width: 640px) 80px, 96px"
+              src={course.imageUrl}
+            />
+          </MediaCardImage>
+        ) : (
+          <MediaCardIcon>
+            <NotebookPenIcon
+              aria-hidden="true"
+              className="size-8 text-muted-foreground/80"
+            />
+          </MediaCardIcon>
+        )}
+
+        <MediaCardContent>
+          <MediaCardHeader>
+            <MediaCardTitle>{course.title}</MediaCardTitle>
+            <MediaCardIndicator />
+          </MediaCardHeader>
+          <MediaCardDescription>{course.description}</MediaCardDescription>
+        </MediaCardContent>
+      </MediaCardTrigger>
+
+      <MediaCardPopover>
+        {isAIGenerated && (
+          <MediaCardPopoverAIWarning className="mb-3">
+            {t(
+              "This content was generated by AI. It may contain errors or inaccuracies and should not be considered professional advice.",
+            )}
+          </MediaCardPopoverAIWarning>
+        )}
+
+        <MediaCardPopoverText>{course.description}</MediaCardPopoverText>
+
+        <MediaCardPopoverMeta>
+          <MediaCardPopoverSource>
+            {course.organization.name}
+          </MediaCardPopoverSource>
+
+          {displayCategories.length > 0 && (
+            <MediaCardPopoverBadges>
+              {displayCategories.map((cat) => (
+                <Badge
+                  key={cat.key}
+                  render={<Link href={`/courses/${cat.key}`} />}
+                  variant="outline"
+                >
+                  {cat.label}
+                </Badge>
+              ))}
+            </MediaCardPopoverBadges>
+          )}
+        </MediaCardPopoverMeta>
+      </MediaCardPopover>
+    </MediaCard>
+  );
+}
