+++ apps/main/src/app/[locale]/(catalog)/b/[brandSlug]/c/[courseSlug]/loading.tsx
@@ -1,0 +1,5 @@
+import { FullPageLoading } from "@zoonk/ui/components/loading";
+
+export default function Loading() {
+  return <FullPageLoading />;
+}
