+++ apps/main/src/app/[locale]/(catalog)/b/[brandSlug]/c/[courseSlug]/page.tsx
@@ -1,38 +1,15 @@
 "use cache";
 
-import { Badge } from "@zoonk/ui/components/badge";
-import {
-  MediaCard,
-  MediaCardContent,
-  MediaCardDescription,
-  MediaCardHeader,
-  MediaCardIcon,
-  MediaCardImage,
-  MediaCardIndicator,
-  MediaCardPopover,
-  MediaCardPopoverAIWarning,
-  MediaCardPopoverBadges,
-  MediaCardPopoverMeta,
-  MediaCardPopoverSource,
-  MediaCardPopoverText,
-  MediaCardTitle,
-  MediaCardTrigger,
-} from "@zoonk/ui/components/media-card";
+import { prisma } from "@zoonk/db";
 import { cacheTagCourse } from "@zoonk/utils/cache";
-import { AI_ORG_SLUG } from "@zoonk/utils/constants";
-import { NotebookPenIcon } from "lucide-react";
 import type { Metadata } from "next";
 import { cacheTag } from "next/cache";
-import Image from "next/image";
 import { notFound } from "next/navigation";
-import { getExtracted, setRequestLocale } from "next-intl/server";
+import { setRequestLocale } from "next-intl/server";
 import { getCourse } from "@/data/courses/get-course";
-import { Link } from "@/i18n/navigation";
-import { getCategories } from "@/lib/categories";
+import { CourseHeader } from "./course-header";
 
 export async function generateStaticParams() {
-  const { prisma } = await import("@zoonk/db");
-
   const course = await prisma.course.findFirst({
     select: { language: true, slug: true },
     where: {
@@ -46,11 +23,7 @@ export async function generateStaticParams() {
   }
 
   return [
-    {
-      brandSlug: "ai",
-      courseSlug: course.slug,
-      locale: course.language,
-    },
+    { brandSlug: "ai", courseSlug: course.slug, locale: course.language },
   ];
 }
 
@@ -58,21 +31,10 @@ export async function generateMetadata({
   params,
 }: PageProps<"/[locale]/b/[brandSlug]/c/[courseSlug]">): Promise<Metadata> {
   const { brandSlug, courseSlug, locale } = await params;
-  const t = await getExtracted({ locale });
-
-  const course = await getCourse({
-    brandSlug,
-    courseSlug,
-    language: locale,
-  });
+  const course = await getCourse({ brandSlug, courseSlug, language: locale });
 
   if (!course) {
-    return {
-      description: t(
-        "Learn with interactive lessons and activities. Explore this course and start your learning journey.",
-      ),
-      title: t("Course not found"),
-    };
+    return {};
   }
 
   return {
@@ -85,92 +47,18 @@ export default async function CoursePage({
   params,
 }: PageProps<"/[locale]/b/[brandSlug]/c/[courseSlug]">) {
   const { brandSlug, courseSlug, locale } = await params;
+  const course = await getCourse({ brandSlug, courseSlug, language: locale });
 
   setRequestLocale(locale);
   cacheTag(cacheTagCourse({ courseSlug }));
-
-  const course = await getCourse({
-    brandSlug,
-    courseSlug,
-    language: locale,
-  });
 
   if (!course) {
     notFound();
   }
 
-  const t = await getExtracted({ locale });
-  const categoryLabels = await getCategories({ locale });
-  const courseCategoryKeys = course.categories.map((c) => c.category);
-  const displayCategories = categoryLabels.filter((cat) =>
-    courseCategoryKeys.includes(cat.key),
-  );
-  const isAIGenerated = brandSlug === AI_ORG_SLUG;
-
   return (
     <main className="flex flex-1 flex-col">
-      <MediaCard>
-        <MediaCardTrigger>
-          {course.imageUrl ? (
-            <MediaCardImage>
-              <Image
-                alt={course.title}
-                className="size-full object-cover"
-                fill
-                sizes="(max-width: 640px) 80px, 96px"
-                src={course.imageUrl}
-              />
-            </MediaCardImage>
-          ) : (
-            <MediaCardIcon>
-              <NotebookPenIcon
-                aria-hidden="true"
-                className="size-8 text-muted-foreground/80"
-              />
-            </MediaCardIcon>
-          )}
-
-          <MediaCardContent>
-            <MediaCardHeader>
-              <MediaCardTitle>{course.title}</MediaCardTitle>
-              <MediaCardIndicator />
-            </MediaCardHeader>
-            <MediaCardDescription>{course.description}</MediaCardDescription>
-          </MediaCardContent>
-        </MediaCardTrigger>
-
-        <MediaCardPopover>
-          {isAIGenerated && (
-            <MediaCardPopoverAIWarning className="mb-3">
-              {t(
-                "This content was generated by AI. It may contain errors or inaccuracies and should not be considered professional advice.",
-              )}
-            </MediaCardPopoverAIWarning>
-          )}
-
-          <MediaCardPopoverText>{course.description}</MediaCardPopoverText>
-
-          <MediaCardPopoverMeta>
-            <MediaCardPopoverSource>
-              {course.organization.name}
-            </MediaCardPopoverSource>
-
-            {displayCategories.length > 0 && (
-              <MediaCardPopoverBadges>
-                {displayCategories.map((cat) => (
-                  <Badge
-                    key={cat.key}
-                    render={<Link href={`/courses/${cat.key}`} />}
-                    variant="outline"
-                  >
-                    {cat.label}
-                  </Badge>
-                ))}
-              </MediaCardPopoverBadges>
-            )}
-          </MediaCardPopoverMeta>
-        </MediaCardPopover>
-      </MediaCard>
+      <CourseHeader brandSlug={brandSlug} course={course} />
     </main>
   );
 }
